issue fixed: 22-signup-survey-submit-button-inside-current-unit-component
